requirements.txt
matplotlib==2.2.3
numpy==1.15.1
pygame==1.9.4
pyserial==3.4
